tast-tests: Create virtual keyboard while in tablet mode.
Update input.Keyboard to look for a tablet mode switch and
create a virtual keyboard instead of looking for a physical
one if we appear to be in tablet mode.

Also update code generator to include _MAX constants and
make little-endian assumptions in the code clearer.

BUG=chromium:930568
TEST=manual: example.Keyboard creates virtual keyboard when
     run against dru but uses physical keyboard when run
     against clamshells or convertibles in laptop mode

Change-Id: Icb7fb71123bef008bd7f99e567e60b0dcd9595fc
Reviewed-on: https://chromium-review.googlesource.com/1549656
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>