Update patch set 1
Patch Set 1:

Neither of those will cover everything, since it's often set in the common.config for architectures in our kernel repos as well.

It actually seems like the static test here would be "is the kernel 3.14 or newer" but I'm not sure if it's enabled for every board for which that is the case. It could also repurpose the ARC software dep since it needs to be present for every board that we enable ARC on.

Patch-set: 1
Reviewer: Gerrit User 1117055 <1117055@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>