Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Since it is complicated to differentiate and have an up to date inventory of boards without display support, the current test does this:
> 
>         result = test_init();
>         if (result == ENODISPLAY) {
>                 printf("[  PASSED  ] graphics_Gbm test no connected display found\n");
>                 return EXIT_SUCCESS;
>         } else if (!result) {
>                 printf("[  FAILED  ] graphics_Gbm test initialization failed\n");
>                 return EXIT_FAILURE;
>         }
> 
> In other words, pass if there is no display. Could we do the same here?

we generally avoid doing this since it produces misleading results -- the test didn't pass in this case; it was skipped since it had unsatisfied dependencies.

shapiroc@ is working on a better way to store devices' characteristics so we can make intelligent decisions when scheduling tests, but in the meantime, we've been relying on the approach used here.

Patch-set: 1
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>