Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > we generally avoid doing this since it produces misleading results -- the test didn't pass in this case; it was skipped since it had unsatisfied dependencies.
> 
> The test passed.

the test didn't actually test the thing that it claims that it's testing. 

> > shapiroc@ is working on a better way to store devices' characteristics so we can make intelligent decisions when scheduling tests, but in the meantime, we've been relying on the approach used here.
> 
> So the problem here is that the graphics team likes to have their tests self contained and not distributed over a huge (infra) surface.

infra is ultimately responsible for executing these tests in the lab. if tests don't expose their dependencies to the scheduling system, then we can't be sure that they're actually verifying correct behavior. this test may be scheduled on a device without a display, where it won't be able to prevent regressions.

Patch-set: 1
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>