Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > Since it is complicated to differentiate and have an up to date inventory of boards without display support, the current test does this:
> > 
> >         result = test_init();
> >         if (result == ENODISPLAY) {
> >                 printf("[  PASSED  ] graphics_Gbm test no connected display found\n");
> >                 return EXIT_SUCCESS;
> >         } else if (!result) {
> >                 printf("[  FAILED  ] graphics_Gbm test initialization failed\n");
> >                 return EXIT_FAILURE;
> >         }
> > 
> > In other words, pass if there is no display. Could we do the same here?
> 
> we generally avoid doing this since it produces misleading results -- the test didn't pass in this case; it was skipped since it had unsatisfied dependencies.


I do not want to have to maintain lists of functionality, since they can get very complex (for example the graphics_idle test looks for PSR, or GPU clocks, or FBC, and other things which are platform-specific; I don't want to start having another set of lists for that).

Doing it this way also prevents me from modifying graphics_gbm in the future to run *some* of the subtests. Right now it's all or nothing, but some (most) of the subtests can definitely run without a monitor. I think we can agree there is value in doing that.

If we have a way to report that some tests were skipped because some dependencies aren't satisfied, that would work very well for this case. Is this something we can do?

> 
> shapiroc@ is working on a better way to store devices' characteristics so we can make intelligent decisions when scheduling tests, but in the meantime, we've been relying on the approach used here.

Patch-set: 1
Reviewer: Gerrit User 1000313 <1000313@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>