arc.ChromeVirtualKeyboard: Check before next input
Previously, arc.ChromeVirtualKeyboard used fixed sleep time to
wait between each keystroke. As Chrome OS VK is written in JavaScript,
it has no guarantee of event ordering, and sometimes it resulted in
flaky test results.

In this CL, each time the test presses a VK key, it waits for the
expected text to appear using UiAutomator, instead of using the fixed
sleep time.

BUG=chromium:945729
TEST=tast run eve-DUT arc.ChromeVirtualKeyboard

Change-Id: I8bc452c66ec26c22d63283027357a7ff27c21a89