tast-tests: do not test /home/root/<uid-hash> for file contexts
Shill is creating directories into it even it's not a bind-mount. The
created file is useless in actual world, and create flakiness for
testing.

This is a quick fix to comment out testing of /home/root/<uid-hash>
since the test should only care the situation when the user is logged in
and it is a bind-mount.

Before confining shill to prevent such file being created, it's better
to skip such files to prevent flakiness that blocks CQ.

BUG=b:130011394,chromium:950317
TEST=tast run

Change-Id: I67a27778cd6e5c77b0003b5bdb33b5af8e92ab5f