tast: Enable ui.VirtualKeyboardSuggestion.
Previously we tried enabling suggestions but it failed on certain devices
in the lab as they had Japanese hardware keyboards, so the input language
defaults to "Japanese Keyboard" input method, which doesn't have a decoder.

Test looks green on informational devices:
https://tastboard.corp.google.com/test/ui.VirtualKeyboardSuggestions

but we will monitor the CQ to ensure that it doesn't fail on Japanese devices.

BUG=chromium:879073
TEST=veyron_minnie, random Japanese device in the lab

Change-Id: Ica3d7390688a0074cce4b3112aa29f77faf46ca8