tast-tests: Fix empty PID file race in security.Minijail.
86517da5b2fc updated security.Minijail's pid-file.static
subtest to loop until it successfully reads the PID file
written by minijail0, but minijail0 doesn't atomically
create the file, so it looks like the subtest is likely
still failing due to performing the read before the contents
have been written. Move the file-content check within the
loop condition so we wait until we get the expected data.

Also map /usr/bin into the chroot since it's needed for the
"sleep" command to work.

BUG=chromium:950504,chromium:949357
TEST=manual: test passes locally; also ensured that "sleep"
     works within the shell

Change-Id: I2691108795ab8a9a3403c83e8d70749ee435d85e