Update patch set 7
Patch Set 7: Verified+1

(10 comments)

Ready for review.

My plan is to add this, then add a follow-up CL which does two things:

1. it organizes and prints various parameters calculated by memoryEqualizingAmount;
2. it special-cases preallocation amount and size of initial tab set for the 8 and 16GB devices.

The calculations in memoryEqualizingAmount are useful to get a picture of what goes on in different devices.  However, in order for the test to run meaningfully in the large devices, the initial tab set has to increase, and this makes the preallocation calculations less useful.

Patch-set: 7
Reviewer: Gerrit User 1000028 <1000028@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1