arc: Run the ADB local server in a new session.
After running ARC tests such as arc.Boot, we should be able to manually
run adb commands to inspect the Android container state.

We used to be able to do this, but today we can't because adb server is
always killed as a side effect of the fix to crbug.com/945626.

Also updates a comment of arc.Command; thanks to the recent
android-tools uprev (crbug.com/737840), exit codes and stdout/stderr
are now correctly handled.

BUG=chromium:954226
TEST=tast run DUT arc.Boot; ssh DUT adb shell echo hi

Change-Id: Idfe183812e15a8d2efc4f9645aab1e8984091cb1