Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > can you file a bug with the reason for this test? have there ever been regressions in lspci on chrome os devices?
> 
> Hi Dan,
> There is a bug https://bugs.chromium.org/p/chromium/issues/detail?id=221100 for lspci but it was created on old os version and we are using this test in our internal testing.
> 
> Do I need to create new bug for it?

thanks. i want to make sure that this test is likely to actually catch regressions, as there's a cost in maintaining tests even while they're passing. (for example, we could write a test for 'ls', but we're not going to because it's unlikely that it would be broken.)

do you have more details about what caused issue 221100 and how it was ultimately fixed? i'm asking because i've never heard of lspci failing. is the concern that we'll fail to compile PCI support into the kernel, or something else? would the problem manifest itself in other ways that would already be caught by other tests?

Patch-set: 2
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>