Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 4:
> > > 
> > > > Patch Set 4:
> > > > 
> > > > > Patch Set 4:
> > > > > 
> > > > > (12 comments)
> > > > > 
> > > > > can you split this change into smaller pieces? it feels like it's doing many not-obviously-to-me-related things, e.g. reading additional stats, logging window dimensions, etc.
> > > > 
> > > > 
> > > > Correct, these are unrelated things, but what are the benefit of separate CLs?  They would likely have dependencies, so I won't be able to, say, pull one out, even if I wanted to.
> > > > 
> > > > Is it because they make the overall review easier?
> > > 
> > > A related observation: the initial CL for such test has a lot of unrelated functions, but I don't have to trickle them in as multiple CLs.
> > 
> > And, there's the issue of testing every CL on various platforms.  I don't suppose it would be kosher to commit intermediate CLs that break the test.
> 
> Did you know mempressure.go is the longest code in the tast-tests repository, even before you add 100+ lines in this CL?
> 
>  % wc -l src/chromiumos/tast/local/bundles/cros/**/*.go | sort -rn
>   35754 total
>    1073 src/chromiumos/tast/local/bundles/cros/platform/mempressure/mempressure.go
>     589 src/chromiumos/tast/local/bundles/cros/security/sandboxed_services.go
>     524 src/chromiumos/tast/local/bundles/cros/vm/crostini_input_latency.go
>     474 src/chromiumos/tast/local/bundles/cros/arc/pip.go
>     411 src/chromiumos/tast/local/bundles/cros/camera/hal3/hal3.go
>  ...
> 
> Separating a large CL into smaller ones often forces you to refactor the code into smaller components, resulting in maintainable and reader-friendly code.
> 
> I'd recommend refactoring this test soon.

I thank you for the advice, but I'll have to do this later.
I am in the middle of breaking the CL into smaller ones, but without refactoring.
The reason is that I don't want to force a complete re-review.  After the split,
the final result will be the same as what is currently under review.
I can refactor after this round of reviews.

I personally didn't see much opportunity for refactoring.  The test is long because it goes through a bunch of phases and performs a number of different actions.  Refactoring would mostly split the test into multiple files, without helping the logic.  I am not sure that this would be much help.  But I will certainly look into it.

Patch-set: 5
Reviewer: Gerrit User 1000028 <1000028@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>