platform.MemoryPressure: measure switch times to cold tabs
Instead of trying to scale the "initial working set" of tabs
to a size that's large enough to induce swapping in systems
with large RAM (8 GiB or more), measure tab switch times to
the LRU tabs after the first discard.  This involves trying
to distinguish discarded and crashed tabs.

A broken devtools connection means that the tab has either
crashed or discarded.  After we detect the first discard
(via the chrome dbus signal) we assume that a broken connection
is due to the discard.  This can have false negatives but they
should be relatively rare.  Before the first discard, broken
connections cause fatal errors.

BUG=none
TEST=ran the test on 2, 4, 8, 16 GiB platforms

Change-Id: I3a404a37b6045ae3932a7e1d4c9fa1f147632b54
Reviewed-on: https://chromium-review.googlesource.com/1576897
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>