Update patch set 2
Patch Set 2: Code-Review+2

were you planning to turn this on in this change, or is it just meant to be manually-set? if the latter, i'd recommend making that clearer in the commit message since it currently sounds like we always use a large screen when there's no display.

Patch-set: 2
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2