Update patch set 4
Patch Set 4:

I am working on Brian's review comment about comparing the number of tab discards with the number of missing tabs.  This is not a small change.  Should it go into a follow-up change, or is the test not usable without it?

Patch-set: 4
Reviewer: Gerrit User 1000028 <1000028@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>