Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I am working on Brian's review comment about comparing the number of tab discards with the number of missing tabs.  This is not a small change.  Should it go into a follow-up change, or is the test not usable without it?

is this a question for me, or for brian or sonny? :-) how not-small is the additional code needed for that?

is this change ready for review now, by the way? it's not clear from me from the earlier comment ("the next patchset will address the new comments", but the next two look like they were just rebases).

Patch-set: 4
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>