Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> Previously, you said capture unittest fails on nami family boards. Is it fixed now?

We haven't seen nami's failures in recent 20 builds.
https://cros-health-prod.googleplex.com/test/video.CaptureUnittests

It might be fixed by another CL as a side effect, but I'm not sure because we haven't found the real cause of that flakiness:(

Anyway, I realized that we have to wait for another 20> builds to check if this test is stable on betty board after CL:1577526.
So, let's keep monitoring.
If this test is still healthy on nami, we may be able to say that the flakiness disappear magically.

Patch-set: 5
Reviewer: Gerrit User 1273836 <1273836@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false