Set device lowmem margin during arc.LowMemoryKiller
Sets /sys/kernel/mm/chromeos-low_mem/margin during test
so that we are sure to consume memory above the threshold
during the test. This prevents the test from quitting too
early and reporting "Nothing got killed". Also detects if
there was a kernel OOM kill, and fails the test in that case.

BUG=chromium:951001
TEST=Run arc.LowMemoryKiller tast test
Change-Id: I2119731d1db42e8b5601a624bdd049c8b28447cd