Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > commit message: "This test always causes kernel panic on some boards, preventing other tests from running."
> > 
> > Lets say we have a test that passes everywhere.
> > Then there is a change that makes daisy reboot.
> > Will the test be removed from testing all other board?
> 
> this is https://crbug.com/933110.
> 
> > Can we keep testing daisy so we keep the repro case, track it and get reminded it is in bad shape?
> 
> i think that we should file an issue with the appropriate priority and fix the bug at that point.

Generally, we don't want to keep tests failing for long time, even if they're just failures and not reboots or kernel crashes, and also even if they're informational. It's a severe cultural problem that leads us to dismiss real failures.

In the future, we'll be able to disable tests per-board basis with crbug.com/933110, but it's not coming soon.

Also IIUC the autotest equivalent to this test is not running on bvt-perbuild, so this is on per with the state of the art.

Patch-set: 2
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>