tast: fix Assistant set volume test failure.
Recently there is a continuous test failure in ui.AssistantTextQueries
with an error msg:

  System volume 75 doesn't match the requested volume 25.

This is caused by a server side change on volume control, where they narrowed
down the volume range from 1-100 to 1-10. Thus, when user try to set the volume
with an absolute value beyond 10 (e.g. "Volume 25"), the assistant will punt
an out-of-range error message.

In this change, we instead set the volume by percentage to be more straightforward
and avoid confusion.

BUG=None.
TEST=tast -verbose run $(DUT) ui.AssistantTextQueries

Change-Id: I7a7e527fd41f49a72c7414311989253502bafe31
Reviewed-on: https://chromium-review.googlesource.com/1579742
Commit-Ready: Meilin Wang <meilinw@chromium.org>
Tested-by: Meilin Wang <meilinw@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>