tast: fix Assistant set volume test failure.
Recently there is a continuous test failure in ui.AssistantTextQueries
with an error msg:
  System volume 75 doesn't match the requested volume 25.

This is caused by a change of rules in volume control, where the max
volume level is now 10, and thus trying to change the volume by setting
a specific level greater than 10 (e.g. "Volume 25") will run into an
out-of-range error and cause test failure.

In this change, we instead set the volume by percentage to avoid confusion
and fit into the new rule.

BUG=None.
TEST=tast -verbose run $(DUT) ui.AssistantTextQueries

Change-Id: I7a7e527fd41f49a72c7414311989253502bafe31