arc.ChromeVirtualKeyboard: Disable shift on failed
It seems after R75-12072.0.0 the behavior of Chrome virtual keyboard
changed and now the virtual keyboard might have shift enabled at the
beggining of the test.

Because of that, the first letter "h" might not be found on the keyboard
as there's only "H".

To workaround that, the test now presses shift key when "h" cannot be
found.

TEST=tast run bob-DUT arc.ChromeVirtualKeyboard
BUG=none
Change-Id: I49a0947fb4f9763df1aaef75da4b33e735e8e707
Reviewed-on: https://chromium-review.googlesource.com/1580840
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>