tast-test: make arc.PIP more reliable
As a temporaty solution, this CL adds a delay before some
activity.WaitForIdle() calls.
This is specially important for the Fling test since
activity.WaitForIdle() might report that the window is idle while it is
being animated, making the test to fail.

Proper solution is being discussed here: http://b/131248000

BUG=chromium:956227
TEST=run "tast $DUT arc.PIP" five times. Passed 5/5.

Change-Id: I8294a707dc75356f13043da6110324e2503fec2b