platform.MemoryPressure: improve tab crash detection
This removes the discard detection via D-Bus.  That was a mistake:
that information is already available via chrome.tabs.query and
is more precise.  When communication with a DevTools target is lost,
use chrome.tabs.query to check if the tab was discarded.  In that
case do NOT return an error, since it is normal behavior.

BUG=none
TEST=ran the test with crashes

Change-Id: Icb6595062a6db43a60a77c27ab759cc4154075af