Update patch set 1
Patch Set 1:

I assume it's just insanely unlikely.

If we wanted to be sure we could have it do a single longer computation that occurs a few stack frames deep, and then as long as the kernel sampling rate is high enough to catch it, we should get a stack. But if the current way works then that's also fine.

Patch-set: 1
Reviewer: Gerrit User 1117055 <1117055@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>