Diagnostics test improvements
Start the deamon before running tests and only test that the urandom
routine passes. We do not want to run all of the routines because this
is an end to end test between the diag tool and the wilco_dtc_supportd
and failures of these routines could happen.

BUG=chromium:919151
TEST='tast -verbose run ${REMOTE_IP} platform.Diagnostics'
Cq-Depend: chromium:1592273

Change-Id: Ia5b29b70a22aeaa1badc6d33851c2abd23e84fbd