tast-tests: Restart ui job at beginning of security.OpenFDs.
Restart the ui job at the beginning of the security.OpenFDs
test. I'm hopeful that this will eliminate flakiness that
appears to be caused by Native-Client-related FDs that might
be getting opened when an earlier test logs in: similar
failures aren't being seen in the Autotest version of this
test, which I think typically runs in a logged-out
environment.

BUG=chromium:924893
TEST=ran it

Change-Id: I1894caeb83f359153ba1fd45bc6bab5cfd966256