tast-tests: stop test after finding first error in example.Keyboard.
The test sometimes fails by not finding expected string for the first
half of the test. It can be either caused by not routing key events to
the input, or not getting the text from the input element correctly.
We're attempting to determine it by capturing screenshot right after it.

BUG=chromium:961173
TEST=none

Change-Id: I1de18b17dfe19a2846dac26ddec16c2fee0a6887