audio.Microphone: Set internal mic active if no active node selected
Sometime the test fails because of broken chrome UI. If chrome is
broken, it can not choose active nodes for us. However, microphone
test should not be affected by this situation.

To fix it, we can set internal mic as a active node if it is not set
automatically.

BUG=chromium:947917
TEST=tast -verbose run $DUT_IP audio.Microphone
The test still pass even if chrome is closed.

Change-Id: I127c1fde70209c2d1cf9940fb1d566e8c1295c9d
Reviewed-on: https://chromium-review.googlesource.com/1605668
Commit-Ready: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>