platform.MemoryPressure: blame more errors on discards
I incorrectly assumed that attempting to execute JS on a
discarded tab would always result in a timeout.  We have evidence
that other errors are possible, so we assume that if the tab
is discarded, any error is due to the discard.  This is not
perfect but is reasonable.

BUG=chromium:957811
TEST=ran the test

Change-Id: Ia6f61800927e567b07ab1c3e861e2b85468aab04
Reviewed-on: https://chromium-review.googlesource.com/1606604
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>