tast-tests: Update meta.RunTests to run meta.LocalVars.
Update the meta.RunTests test to execute a new
meta.LocalVars test to verify the end-to-end functionality
of the new -var flag to "tast run".

meta.RunTests passes a "meta" test variable at runtime.
meta.LocalVars writes the value of the variable to an output
file, which meta.RunTests later inspects.

BUG=chromium:942808
TEST=ran the test (after much emerging and symlinking; meta
     tests are hard to run)

Cq-Depend: chromium:1626177
Change-Id: I5d690e4f0a102d9d28578231d32d0d0ee1508edd