Update patch set 2
Patch Set 2:

> Patch Set 2: Verified+1
> 
> I am still not 100% confident if this is the actual cause of the flakiness, but I think this would be somewhat reasonable to fix.
> I once observed that document.activeElement evaluated to null inside the test function, by repeatedly running the test. (and the UI showed up normally at least after the test ended.)

Ah, this is a nice finding! So essentially we want to elementExpr to evaluate to non-null because it's the case before the document is fully loaded.

This is an example test so we want to fix the issue in somewhat generic way. What do you think about instead waiting for the page load by polling "document.readyState === 'complete'"?

https://cs.corp.google.com/chromeos_public/src/platform/tast-tests/src/chromiumos/tast/local/bundles/cros/ui/chrome_login.go?q=file:tast-tests+complete&g=0&l=67

Patch-set: 2
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>