tast-test: improve PIP reliability and add helper functions (1/2)
This CL (part 1 of 2) improves arc.PIP test reliability by using a
better way to detech whether the shelf is at the bottom.

It also adds helper functions needed to manipulate the shelf. These
functions will be used in the next CL (part 2).

This CL is a subset of this one:  https://crrev.com/c/1577899

BUG=chromium:9433850
TEST=tast -verbose run cros0 arc.Pip

Change-Id: Ib0c3d03c5b17ec80439f12b30221fd6341ec2de8