tast-test: improve reliability in arc.ResizeActivity test
This CL improves reliability in arc.ResizeActivity test.
This test is failing on some boards with:
Failed to resize window: failed to start the swipe gesture: coordinates (2303, 1986) outside valid bounds (3520, 1984)

Meaning that the test is trying to inject a touch outside the display bounds.

This is being caused both by:
- The test doesn't wait until the app is in freeform mode
- Activity is adding pixels to "src" but they could be outside the
display.

This CL fixes those two conditions. And adds some cosmetic fixes in the
code.

BUG=chromium:966210
TEST=tast -verbose run cros0 arc.ResizeActivity
     tested on Nocture, Kevin and Eve.

Change-Id: Ic15a4537555348ca7a48f7cec1dcde9ed9575220