tast-tests: vm.CrostiniCPUPerf: skip sysbench if not available
Arm builds don't currently have sysbench (https://crbug.com/922178), and
this doesn't look easy to resolve.  In the interim, skip the sysbench
portion of the test when sysbench isn't available so we can at least get
lmbench results for arm platforms.

While we're cleaning this up, split the sysbench and lmbench tests into
functions to make the overall flow of the test easier to understand.

BUG=chromium:922178
TEST=tast run kevin vm.CrostiniCPUPerf

Change-Id: I1eb9333708c3ed79258d194e8dbea9691ee19bdb
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>