tast-test: add virtual touch support
This CL creates a virtual touchscreen device when a physical toushcreen
device is not found.

In theory tests that requires a touchscreen should not be run on devices
without one. But since we are using unibuilds test that require a
touchscreen are being run on devices without one (see crbug/950346).

This CL mitigates that issue by creating a virtual touchscreen (using
the same API as the virtual keyboard).

BUG=chromium:944621
TEST=tast -verbose run cros0 example.Touch in Grunt
     Also forced to used Virtual Touch in Eve and Nocturne and works Ok
     too.
TEST=tast -verbose run cros0 example.Keyboard forcing to use virtual
     keyboard (since this CL makes some cosmetic changes in keyboard)
Change-Id: I2c0c5830aa0660b6ae50872c2c5011ec7816b419