video: log each polled value of currentTime().
Right now, we use Conn.WaitForExpr() to wait for currentTime() > 0.9.
However, if it failes to poll, we don't know the value of currentTime().
So it uses testing.Poll() to poll currentTime() and logs the value for further debugging.

BUG=chromium:963331, chromium:963334
TEST=tast run $IP video.PlayDecodeAccelUsed.*

Change-Id: I09e138c2e9d48bc92458fc2523493a46f880df50
Reviewed-on: https://chromium-review.googlesource.com/1627412
Commit-Ready: Shuo-Peng Liao <deanliao@chromium.org>
Tested-by: Shuo-Peng Liao <deanliao@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>