Update patch set 2
Patch Set 2:

Thanks for the example!

Polling "currentTime < 0.9" should finish in a second ideally, so the error suggests that the test used most of its time on polling currentTime, not in operations before that. (Of course it's possible that the operations before polling took 2m59s and the polling was aborted due to 3m timeout, but it sounds very unlikely.) Then I'm afraid this logging does not help much for debugging the issue.


Generally we ask test authors to be careful about leaving useful logs.

https://chromium.googlesource.com/chromiumos/platform/tast/+/HEAD/docs/writing_tests.md#When-to-log

That's a lesson we learned from CrOS autotest where we left thousands of lines of logs for every test and made debugging harder.


So I recommend:

- Set the explicit timeout for polling currentTime so we don't continue polling until we hit the test timeout. Actually I see you do this in crrev.com/c/1627412.
- When polling times out, run some diagnosis logic to gather debug information.

Patch-set: 2
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>