tast-tests: network.ConfigureServiceForProfile: grab check_ethernet lock
This script stops/starts shill a few times, to reset state and to test
that its own state remains across restart. This causes us to lose
connectivity briefly, and it gives the recover_duts/check_ethernet.hook
infrastructure a chance to start screwing us up. This is particularly
problematic if it:

 * decides to start disabling/re-enabling USB ethernet at the wrong
   time, as that may prevent our Ethernet Service (that this test looks
   for) from showing up promptly; or
 * restarts shill, which will cause us to lose our D-Bus connection and
   time out

Solution: hold the check_ethernet lock for the duration of this test.

See also http://crrev.com/c/1637489, which improves the locking on the
recover_duts side.

Note: even with this lock usage, there is still potentially some ramp-up
time after this test, where we are no longer holding the lock but may
have brief connectivity loss (while Ethernet is getting re-configured).
That's a problem for the next test though ;)

Note: shill.Safe{Stop,Start}() utilize the old
/run/lock/shill-start.lock properly, but unfortunately this is all
pretty racy too, since:
 (*) check_ethernet.hook doesn't currently acquire the lock -- it only
     reads it -- so there's still a race window
 (*) even if shill has started (in the Upstart sense), it may not have
     finished configuring Ethernet; so check_ethernet.hook may again
     find the lock "open", but connectivity not in good shape yet
This is a problem for another day, requiring total
revamp/replacement/removal of shill-start.lock. Fortunately,
'autotest_pause_ethernet_hook' is a sufficient superset of that lock,
and it satisfies the needs of this test.

I also add the clean-profile-on-exit logic that is present in the
Autotest version, but not the Tast version of this test.

BUG=chromium:966965
TEST=`touch /mnt/stateful_partition/.labmachine; start recover_duts`,
     run network.ConfigureServiceForProfile repeatedly

Change-Id: I84a645591dff248547fd3d445ed89055beb83ea5